check des connexion et stmt a null avant de les close. check si la connexion a la DB a pu se faire sinon renvoie un objet JSON {"Success": "KO"} check si le config.resources a pu etre ouvert avant de load les properties.

+++ RSS-Server/src/com/rss/server/DatabaseManager.java
@@ -15,9 +15,9 @@ import org.json.JSONObject;
 
 public class DatabaseManager {
 
-	private String _user;
-	private String _pwd;
-	private String _url;
+	private String _user = "root";
+	private String _pwd = "toor";
+	private String _url = "jdbc:mysql://localhost:3306/rssaggregatordb";
 
 	Properties prop = new Properties();
 	InputStream input = null;
@@ -25,10 +25,13 @@ public class DatabaseManager {
 	public DatabaseManager() throws SQLException, IOException {
 		DriverManager.registerDriver(new com.mysql.jdbc.Driver ());
 		input = getClass().getClassLoader().getResourceAsStream("config.properties");
-		prop.load(input);
-		_user = prop.getProperty("rssAggreg.user");
-		_pwd = prop.getProperty("rssAggreg.pwd");
-		_url = prop.getProperty("rssAggreg.url");
+		if (input != null)
+		{
+			prop.load(input);
+			_user = prop.getProperty("rssAggreg.user");
+			_pwd = prop.getProperty("rssAggreg.pwd");
+			_url = prop.getProperty("rssAggreg.url");
+		}
 	}
 	
 	public String checkUser(String userName, String userPwd, int req) {
@@ -39,6 +42,11 @@ public class DatabaseManager {
 		
 		try {
 			connexion = DriverManager.getConnection(_url, _user, _pwd);
+			if (connexion == null)
+			{
+				response.put("Success", "KO");
+				return response.toString();
+			}
 			String sqlQuery = "SELECT * from user WHERE login LIKE ? AND password LIKE ?";
 			stmt = connexion.prepareStatement(sqlQuery);
 			stmt.setString(1, userName);
@@ -54,7 +62,6 @@ public class DatabaseManager {
 				return (i > 0 ? "KO" : "OK");
 			else if (req == 0 && i == 0)
 			{
-				
 				response.put("Success", "KO");
 				return response.toString();
 			}
@@ -69,8 +76,10 @@ public class DatabaseManager {
 		}
 		finally {
 			try {
-				stmt.close();
-				connexion.close();
+				if (stmt != null)
+					stmt.close();
+				if (connexion != null)
+					connexion.close();
 			} catch (SQLException ignore) {
 			}
 		}
@@ -101,10 +110,16 @@ public class DatabaseManager {
 	public String addUser(String userName, String userPwd) {
 		
 		checkUser(userName, userPwd, 1);
+		JSONObject response = new JSONObject();
 		Connection connexion = null;
 		PreparedStatement stmt = null;
 		try {
 			connexion = DriverManager.getConnection(_url, _user, _pwd);
+			if (connexion == null)
+			{
+				response.put("Success", "KO");
+				return response.toString();
+			}
 			String sqlQuery = "INSERT INTO user (login, password) VALUES (?, ?)";
 			stmt = connexion.prepareStatement(sqlQuery);
 			stmt.setString(1, userName);
@@ -115,12 +130,13 @@ public class DatabaseManager {
 		}
 		finally {
 			try {
-				stmt.close();
-				connexion.close();
+				if (stmt != null)
+					stmt.close();
+				if (connexion != null)
+					connexion.close();
 			} catch (SQLException ignore) {
 			}
 		}
-		JSONObject response = new JSONObject();
 		response.put("Success", "OK");
 		return response.toString();
 	}
@@ -132,6 +148,11 @@ public class DatabaseManager {
 		JSONObject response = new JSONObject();
 		try {
 			connexion = DriverManager.getConnection(_url, _user, _pwd);
+			if (connexion == null)
+			{
+				response.put("Success", "KO");
+				return response;
+			}
 			String sqlQuery = "SELECT * from rss_domain WHERE title LIKE ? AND link LIKE ?";
 			stmt = connexion.prepareStatement(sqlQuery);
 			stmt.setString(1, rssName);
@@ -152,8 +173,10 @@ public class DatabaseManager {
 		}
 		finally {
 			try {
-				stmt.close();
-				connexion.close();
+				if (stmt != null)
+					stmt.close();
+				if (connexion != null)
+					connexion.close();
 			} catch (SQLException ignore) {
 			}
 		}
@@ -167,10 +190,14 @@ public class DatabaseManager {
 		PreparedStatement stmt = null;
 		String rssId = null;
 //		String response= null;
-		String ret = null;
 		JSONObject response = new JSONObject();
 		try {
 			connexion = DriverManager.getConnection(_url, _user, _pwd);
+			if (connexion == null)
+			{
+				response.put("Success", "KO");
+				return response.toString();
+			}
 			response = checkRSS(rssName, rssURL);
 			if (response.get("Success").equals("KO"))
 			{
@@ -200,8 +227,10 @@ public class DatabaseManager {
 		}
 		finally {
 			try {
-				stmt.close();
-				connexion.close();
+				if (stmt != null)
+					stmt.close();
+				if (connexion != null)
+					connexion.close();
 			} catch (SQLException ignore) {
 			}
 		}
@@ -213,6 +242,10 @@ public class DatabaseManager {
 		PreparedStatement stmt = null;
 		try {
 			connexion = DriverManager.getConnection(_url, _user, _pwd);
+			if (connexion == null)
+			{
+				return false;
+			}
 			String sqlQuery = "SELECT * FROM user_domain_assoc WHERE user_id = ? AND rss_domain_id = ?";
 			stmt = connexion.prepareStatement(sqlQuery);
 			stmt.setString(1, userId);
@@ -234,6 +267,11 @@ public class DatabaseManager {
 		JSONObject response = new JSONObject();
 		try {
 			connexion = DriverManager.getConnection(_url, _user, _pwd);
+			if (connexion == null)
+			{
+				response.put("Success", "KO");
+				return response.toString();
+			}
 			// remove link between user and rss
 			
 			String sqlQuery = "DELETE FROM user_domain_assoc WHERE user_domain_assoc.user_id = ? AND user_domain_assoc.rss_domain_id = (SELECT id FROM rss_domain WHERE rss_domain.id = ?)";
@@ -249,8 +287,10 @@ public class DatabaseManager {
 		}
 		finally {
 			try {
-				stmt.close();
-				connexion.close();
+				if (stmt != null)
+					stmt.close();
+				if (connexion != null)
+					connexion.close();
 			} catch (SQLException ignore) {
 			}
 		}
